# Optimize cugraph-dgl MFG creation  