--- a/python/cugraph-dgl/examples/dataset_from_disk_cudf.ipynb
+++ b/python/cugraph-dgl/examples/dataset_from_disk_cudf.ipynb
@@ -23,8 +23,9 @@
     "import rmm\n",
     "import torch\n",
     "from rmm.allocators.torch import rmm_torch_allocator\n",
-    "# rmm.reinitialize(pool_allocator = True, initial_pool_size = 15e9, maximum_pool_size=24e9)\n",
-    "# torch.cuda.memory.change_current_allocator(rmm_torch_allocator)"
+    "#rmm.reinitialize(initial_pool_size=15e9)\n",
+    "rmm.mr.set_current_device_resource(rmm.mr.CudaAsyncMemoryResource(initial_pool_size=15e9))\n",
+    "torch.cuda.memory.change_current_allocator(rmm_torch_allocator)"
    ]
   },
   {
@@ -50,7 +51,6 @@
    "source": [
     "def load_dgl_dataset(dataset_name='ogbn-products'):\n",
     "    from ogb.nodeproppred import DglNodePropPredDataset\n",
-    "\n",
     "    dataset_root = '/raid/vjawa/gnn/'\n",
     "    dataset =  DglNodePropPredDataset(name = dataset_name, root=dataset_root)\n",
     "    split_idx = dataset.get_idx_split()\n",
@@ -85,7 +85,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 5,
+   "execution_count": 16,
    "id": "eff3d77b",
    "metadata": {
     "tags": []
@@ -103,7 +103,7 @@
     "    g,                               \n",
     "    train_idx.to('cuda'),                        # train_nid must be on GPU.\n",
     "    sampler,\n",
-    "    sampling_output_dir=temp_dir.name, # Path to save sampling results to\n",
+    "    sampling_output_dir='/raid/vjawa/obgn_products/', # Path to save sampling results to\n",
     "    device=torch.device('cuda'),    # The device argument must be GPU.\n",
     "    num_workers=0,                 # Number of workers must be 0.\n",
     "    batch_size=batch_size,\n",
@@ -115,7 +115,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 6,
+   "execution_count": 17,
    "id": "94003c30-756f-4cdb-856a-dec16a5fb4dc",
    "metadata": {
     "tags": []
@@ -125,14 +125,12 @@
      "name": "stdout",
      "output_type": "stream",
      "text": [
-      "CPU times: user 7.97 s, sys: 8.78 s, total: 16.8 s\n",
-      "Wall time: 18.3 s\n"
+      "6.91 s ± 1.31 s per loop (mean ± std. dev. of 7 runs, 1 loop each)\n"
      ]
     }
    ],
    "source": [
-    "%%time\n",
-    "\n",
+    "%%timeit\n",
     "batch_stats = {}\n",
     "for batch_id,(input_nodes, output_nodes, blocks) in enumerate(dataloader):\n",
     "    batch_stats[batch_id]={'input_nodes':len(input_nodes),'output_nodes':len(output_nodes)}"
@@ -140,7 +138,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 7,
+   "execution_count": 18,
    "id": "d8488e64-ba92-40c6-8e76-3898b1ca4317",
    "metadata": {
     "tags": []
@@ -161,7 +159,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 10,
+   "execution_count": 19,
    "id": "0d147756-6410-4b71-aac1-9ef1e3df8fff",
    "metadata": {
     "tags": []
@@ -174,7 +172,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 11,
+   "execution_count": 9,
    "id": "7cb2cc68-b4ff-43f2-8b12-b2808510b3f2",
    "metadata": {
     "tags": []
@@ -199,7 +197,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 12,
+   "execution_count": 10,
    "id": "7988aca2-7bfb-4200-ac87-008e30c670fb",
    "metadata": {
     "tags": []
@@ -209,17 +207,35 @@
      "name": "stdout",
      "output_type": "stream",
      "text": [
-      "CPU times: user 10min 7s, sys: 1min 18s, total: 11min 26s\n",
-      "Wall time: 16.6 s\n"
+      "5.66 s ± 418 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)\n"
      ]
     }
    ],
    "source": [
-    "%%time\n",
+    "%%timeit\n",
     "dgl_batch_stats = {}\n",
     "for batch_id,(input_nodes, output_nodes, blocks) in enumerate(dataloader):\n",
     "    dgl_batch_stats[batch_id]={'input_nodes':len(input_nodes),'output_nodes':len(output_nodes)}"
    ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 11,
+   "id": "57022ea6-d2fc-4334-a086-82201e8814c8",
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "del dataloader\n",
+    "del g"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": null,
+   "id": "a1fc5ea7-247f-4f4c-a2fa-588bf8f8dfc1",
+   "metadata": {},
+   "outputs": [],
+   "source": []
   }
  ],
  "metadata": {
@@ -238,7 +254,7 @@
    "name": "python",
    "nbconvert_exporter": "python",
    "pygments_lexer": "ipython3",
-   "version": "3.10.9"
+   "version": "3.10.11"
   },
   "vscode": {
    "interpreter": {