--- a/python/cugraph-dgl/examples/dataset_from_disk_cudf.ipynb
+++ b/python/cugraph-dgl/examples/dataset_from_disk_cudf.ipynb
@@ -85,7 +85,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 16,
+   "execution_count": 26,
    "id": "eff3d77b",
    "metadata": {
     "tags": []
@@ -115,20 +115,12 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 17,
+   "execution_count": null,
    "id": "94003c30-756f-4cdb-856a-dec16a5fb4dc",
    "metadata": {
     "tags": []
    },
-   "outputs": [
-    {
-     "name": "stdout",
-     "output_type": "stream",
-     "text": [
-      "6.91 s ± 1.31 s per loop (mean ± std. dev. of 7 runs, 1 loop each)\n"
-     ]
-    }
-   ],
+   "outputs": [],
    "source": [
     "%%timeit\n",
     "batch_stats = {}\n",
@@ -138,7 +130,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 18,
+   "execution_count": null,
    "id": "d8488e64-ba92-40c6-8e76-3898b1ca4317",
    "metadata": {
     "tags": []
@@ -159,7 +151,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 19,
+   "execution_count": null,
    "id": "0d147756-6410-4b71-aac1-9ef1e3df8fff",
    "metadata": {
     "tags": []
@@ -172,7 +164,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 9,
+   "execution_count": null,
    "id": "7cb2cc68-b4ff-43f2-8b12-b2808510b3f2",
    "metadata": {
     "tags": []
@@ -197,20 +189,12 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 10,
+   "execution_count": null,
    "id": "7988aca2-7bfb-4200-ac87-008e30c670fb",
    "metadata": {
     "tags": []
    },
-   "outputs": [
-    {
-     "name": "stdout",
-     "output_type": "stream",
-     "text": [
-      "5.66 s ± 418 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)\n"
-     ]
-    }
-   ],
+   "outputs": [],
    "source": [
     "%%timeit\n",
     "dgl_batch_stats = {}\n",
@@ -220,7 +204,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 11,
+   "execution_count": null,
    "id": "57022ea6-d2fc-4334-a086-82201e8814c8",
    "metadata": {},
    "outputs": [],
@@ -228,14 +212,6 @@
     "del dataloader\n",
     "del g"
    ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": null,
-   "id": "a1fc5ea7-247f-4f4c-a2fa-588bf8f8dfc1",
-   "metadata": {},
-   "outputs": [],
-   "source": []
   }
  ],
  "metadata": {