--- a/python/cugraph-dgl/examples/dataset_from_disk_cudf.ipynb
+++ b/python/cugraph-dgl/examples/dataset_from_disk_cudf.ipynb
@@ -23,8 +23,8 @@
     "import rmm\n",
     "import torch\n",
     "from rmm.allocators.torch import rmm_torch_allocator\n",
-    "#rmm.reinitialize(initial_pool_size=15e9)\n",
-    "rmm.mr.set_current_device_resource(rmm.mr.CudaAsyncMemoryResource(initial_pool_size=15e9))\n",
+    "rmm.reinitialize(initial_pool_size=15e9)\n",
+    "#rmm.mr.set_current_device_resource(rmm.mr.CudaAsyncMemoryResource(initial_pool_size=15e9))\n",
     "torch.cuda.memory.change_current_allocator(rmm_torch_allocator)"
    ]
   },
@@ -85,7 +85,17 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 26,
+   "execution_count": 5,
+   "id": "3808055c-2d7d-4cc7-b1bd-2fe9edd6eb95",
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "!rm -rf \"/raid/vjawa/obgn_products_sampling/\""
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 6,
    "id": "eff3d77b",
    "metadata": {
     "tags": []
@@ -95,7 +105,6 @@
     "g, train_idx = load_dgl_dataset()\n",
     "g = cugraph_dgl.cugraph_storage_from_heterograph(g, single_gpu=single_gpu)\n",
     "\n",
-    "temp_dir = tempfile.TemporaryDirectory()\n",
     "batch_size = 1024\n",
     "fanout_vals=[25, 25]\n",
     "sampler = cugraph_dgl.dataloading.NeighborSampler(fanout_vals)\n",
@@ -103,24 +112,32 @@
     "    g,                               \n",
     "    train_idx.to('cuda'),                        # train_nid must be on GPU.\n",
     "    sampler,\n",
-    "    sampling_output_dir='/raid/vjawa/obgn_products/', # Path to save sampling results to\n",
+    "    sampling_output_dir=\"/raid/vjawa/obgn_products_sampling/\", # Path to save sampling results to\n",
     "    device=torch.device('cuda'),    # The device argument must be GPU.\n",
     "    num_workers=0,                 # Number of workers must be 0.\n",
     "    batch_size=batch_size,\n",
-    "    batches_per_partition=20,\n",
-    "    seeds_per_call=50_000,\n",
+    "    batches_per_partition=50,\n",
+    "    seeds_per_call=50*batch_size,\n",
     "    drop_last=False,\n",
     "    shuffle=False)"
    ]
   },
   {
    "cell_type": "code",
-   "execution_count": null,
+   "execution_count": 7,
    "id": "94003c30-756f-4cdb-856a-dec16a5fb4dc",
    "metadata": {
     "tags": []
    },
-   "outputs": [],
+   "outputs": [
+    {
+     "name": "stdout",
+     "output_type": "stream",
+     "text": [
+      "7.25 s ± 916 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)\n"
+     ]
+    }
+   ],
    "source": [
     "%%timeit\n",
     "batch_stats = {}\n",
@@ -130,7 +147,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": null,
+   "execution_count": 8,
    "id": "d8488e64-ba92-40c6-8e76-3898b1ca4317",
    "metadata": {
     "tags": []
@@ -151,7 +168,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": null,
+   "execution_count": 9,
    "id": "0d147756-6410-4b71-aac1-9ef1e3df8fff",
    "metadata": {
     "tags": []
@@ -164,7 +181,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": null,
+   "execution_count": 10,
    "id": "7cb2cc68-b4ff-43f2-8b12-b2808510b3f2",
    "metadata": {
     "tags": []
@@ -189,12 +206,20 @@
   },
   {
    "cell_type": "code",
-   "execution_count": null,
+   "execution_count": 11,
    "id": "7988aca2-7bfb-4200-ac87-008e30c670fb",
    "metadata": {
     "tags": []
    },
-   "outputs": [],
+   "outputs": [
+    {
+     "name": "stdout",
+     "output_type": "stream",
+     "text": [
+      "4.22 s ± 345 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)\n"
+     ]
+    }
+   ],
    "source": [
     "%%timeit\n",
     "dgl_batch_stats = {}\n",
@@ -204,7 +229,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": null,
+   "execution_count": 12,
    "id": "57022ea6-d2fc-4334-a086-82201e8814c8",
    "metadata": {},
    "outputs": [],