--- a/python/cugraph-dgl/examples/dataset_from_disk_cudf.ipynb
+++ b/python/cugraph-dgl/examples/dataset_from_disk_cudf.ipynb
@@ -24,6 +24,7 @@
     "import torch\n",
     "from rmm.allocators.torch import rmm_torch_allocator\n",
     "rmm.reinitialize(initial_pool_size=15e9)\n",
+    "#Switch to async pool in case of memory issues due to fragmentation of the pool\n", 
     "#rmm.mr.set_current_device_resource(rmm.mr.CudaAsyncMemoryResource(initial_pool_size=15e9))\n",
     "torch.cuda.memory.change_current_allocator(rmm_torch_allocator)"
    ]
@@ -112,7 +113,7 @@
     "    g,                               \n",
     "    train_idx.to('cuda'),                        # train_nid must be on GPU.\n",
     "    sampler,\n",
-    "    sampling_output_dir=\"/raid/vjawa/obgn_products_sampling/\", # Path to save sampling results to\n",
+    "    sampling_output_dir=\"/raid/vjawa/obgn_products_sampling/\", # Path to save sampling results to, Change to the fastest IO path available\n",
     "    device=torch.device('cuda'),    # The device argument must be GPU.\n",
     "    num_workers=0,                 # Number of workers must be 0.\n",
     "    batch_size=batch_size,\n",