This Pull Request significantly optimizes the creation process of Message Flow Graphs (MFG), achieving a **threefold** speed improvement. The enhancements were realized by leveraging the optimizations found by @alexbarghi-nv during bulk sampling.

Furthermore, the PR incorporates the usage of torch.searchsorted as recommended by @alexbarghi-nv.

The impact of these improvements can be seen clearly when sampling on ogbn-products.

**Post-Optimization**

```python3
%%timeit -n 20

sampled_df = cudf.read_parquet("/raid/vjawa/batch=0-19.parquet")
sampled_df.rename(columns={"sources":"destinations",
                   "destinations":"sources"},
                  inplace=True)

mfgs = create_homogeneous_sampled_graphs_from_dataframe(sampled_df, total_number_of_nodes=2449029)
```

This leads to an execution speed of `132 ms ± 1.11 ms per loop (mean ± std. dev. of 7 runs, 20 loops each)`.


**Pre-Optimization**

```python3
%%timeit -n 20

sampled_df = cudf.read_parquet("/raid/vjawa/test_products_dataset_samples/batch=0-19_compressed.parquet")
sampled_df.rename(columns={"sources":"destinations",
                   "destinations":"sources"},
                  inplace=True)

mfgs = create_homogeneous_sampled_graphs_from_dataframe(sampled_df, total_number_of_nodes=2449029)
```

The performance before the changes was `395 ms ± 90 ms per loop (mean ± std. dev. of 7 runs, 20 loops each)`.